Add files via upload
Teclov is an asset management company. Teclov wants to make investments in a few companies. The CEO of Teclov wants to understand the global trends in investments so that she can take the investment decisions effectively. So I have performed data analysis and have tried to solve the questions provided.